% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puddlr.R
\name{CalculatePredictorDispersion}
\alias{CalculatePredictorDispersion}
\title{Calculates dispersion of predictors, using supplied measures.  Defaults to
using median absolute deviation (MAD).}
\usage{
CalculatePredictorDispersion(puddlr, dispersion.measure = "MAD")
}
\arguments{
\item{puddlr}{A puddlr object to calculate dispersion of predictors}

\item{dispersion.measure}{whether to scale predictor matrix. Default
value = 'MAD' (Median Absolute Deviation).}
}
\value{
A puddlr object with dispersion calculated and stored in the
        'predictor.meta' attribute.
}
\description{
Calculates dispersion of predictors, using supplied measures.  Defaults to
using median absolute deviation (MAD).
}
