% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puddlr.R
\name{MaskPredictors}
\alias{MaskPredictors}
\title{Masks normalized predictors from downstream analysis, either by a character
vector of predictor names, or a boolean vector where TRUE indicates that a
predictor should be masked.  Masking predictors will remove all
dimensionality reductions from a puddlr object, and these will need to be
re-computed if required.}
\usage{
MaskPredictors(puddlr, predictors.to.keep)
}
\arguments{
\item{puddlr}{A puddlr object to calculate dispersion of predictors}

\item{predictors.to.keep}{either a character vevtor or boolean vector containing
which predictors to keep. Indexed from the
"predictors" attribute}
}
\value{
A puddlr object with predictors masked. The "raw.predictors"
        attribute will remain unchanged.
}
\description{
Masks normalized predictors from downstream analysis, either by a character
vector of predictor names, or a boolean vector where TRUE indicates that a
predictor should be masked.  Masking predictors will remove all
dimensionality reductions from a puddlr object, and these will need to be
re-computed if required.
}
