% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puddlr.R
\name{NormalizePredictors}
\alias{NormalizePredictors}
\title{Center and normalize predictors for linear dimensionality reduction.
Also prunes all observations where NA values are identified for either
the response or any of the predictors.}
\usage{
NormalizePredictors(puddlr, scale = TRUE, center = TRUE, max.na.ratio = 0.5)
}
\arguments{
\item{puddlr}{A puddlr object to normalize in-place}

\item{scale}{whether to scale predictor matrix}

\item{center}{whether to center predictor matrix}

\item{max.na.ratio}{exclude predictors where the percent of NA observations
exceeds \code{max.na.ratio}. Default = \code{0.5}.}
}
\value{
A normalized puddlr object
}
\description{
Center and normalize predictors for linear dimensionality reduction.
Also prunes all observations where NA values are identified for either
the response or any of the predictors.
}
