% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puddlr.R
\name{rsq.v}
\alias{rsq.v}
\title{forked from rsq package.
Zhang (2017): variance-function-based R-squared.}
\usage{
rsq.v(fitObj, adj = FALSE)
}
\arguments{
\item{fitObj}{glm or lm object to calculate R-squared for}

\item{adj}{whether to adjust R-squared value.}
}
\value{
A normalized puddlr object
}
\description{
forked from rsq package.
Zhang (2017): variance-function-based R-squared.
}
