% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puddlr.R
\name{RunGLM}
\alias{RunGLM}
\title{Runs generalized linear model on a puddlr object.  Intended to be run after
normalization and removal of incomplete observations.}
\usage{
RunGLM(puddlr, formula, family, reduction, n.components, adj.rsq = FALSE)
}
\arguments{
\item{puddlr}{A puddlr object}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be 
fitted. The details of model specification can be found 
under `glm`.  The name of the response variable in the
formula will be used to name the response variable.}

\item{family}{a description of the error distribution and link function
to be used in the model.  Passed to the argument of the
same name under `glm`.}

\item{reduction}{a string specifying the linear dimensionality reduction to
use. Valid options are \code{'pca'}.  Default = \code{'pca'}}

\item{n.components}{an integer specifying the total number of reduction
components to use in the GLM.}

\item{adj.rsq}{boolean flag specifying whether to adjust the pseudo-R^2
goodness-of-fit calculated value.}
}
\value{
puddlr object with glm
}
\description{
Runs generalized linear model on a puddlr object.  Intended to be run after
normalization and removal of incomplete observations.
}
