% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puddlr.R
\name{RunICA}
\alias{RunICA}
\title{Runs ICA for linear dimensionality reduction on the predictors of a 
normalized puddlr object}
\usage{
RunICA(puddlr, nc = 20)
}
\arguments{
\item{puddlr}{A puddlr object}

\item{nc}{number of components to extract. Default = 20}
}
\value{
puddlr object with pca reduction
}
\description{
Runs ICA for linear dimensionality reduction on the predictors of a 
normalized puddlr object
}
