% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puddlr.R
\name{RunPCA}
\alias{RunPCA}
\title{Runs PCA for linear dimensionality reduction on the predictors of a 
normalized puddlr object}
\usage{
RunPCA(puddlr, center = TRUE, scale = TRUE)
}
\arguments{
\item{puddlr}{A puddlr object}

\item{center}{whether to zero-center}

\item{scale}{whether to center data to unit variance}
}
\value{
puddlr object with pca reduction
}
\description{
Runs PCA for linear dimensionality reduction on the predictors of a 
normalized puddlr object
}
