% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puddlr.R
\name{ScanComponentsSubset}
\alias{ScanComponentsSubset}
\title{ScanComponentsSubset}
\usage{
ScanComponentsSubset(
  puddlr,
  n.to.scan.vec,
  k.cross = 7,
  rand.seed = 42,
  formula,
  family,
  reduction,
  adj.rsq
)
}
\arguments{
\item{puddlr}{A puddlr object}

\item{n.to.scan.vec}{Vector of possible values for n.components to scan for
GLM fitting.}

\item{k.cross}{number of train-test splits. (corresponds to k-fold cross
validation).  Default = 7.}

\item{rand.seed}{random seed for cross validation split, default=42.}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be 
fitted. The details of model specification can be found 
under `glm`.  The name of the response variable in the
formula will be used to name the response variable.}

\item{family}{a description of the error distribution and link function
to be used in the model.  Passed to the argument of the
same name under `glm`.}

\item{reduction}{a string specifying the linear dimensionality reduction to
use. Valid options are \code{'pca'}.  Default = \code{'pca'}}

\item{adj.rsq}{boolean flag specifying whether to adjust the pseudo-R^2
goodness-of-fit calculated value.}
}
\value{
puddlr object with scanned n components by rsq plot.
}
\description{
ScanComponentsSubset
}
