% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puddlr.R
\name{CreatePuddlrObject}
\alias{CreatePuddlrObject}
\title{Create Puddlr object}
\usage{
CreatePuddlrObject(response, predictors)
}
\arguments{
\item{response}{vector of observed values for the response variable}

\item{predictors}{matrix of predictors, formatted as observations (rows)
by variables (columns)}
}
\value{
a Puddlr object
}
\description{
Create Puddlr object
}
