% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puddlr.R
\name{RunPCA}
\alias{RunPCA}
\title{Runs PCA for linear dimensionality reduction on the predictors of a 
normalized puddlr object}
\usage{
RunPCA(puddlr)
}
\arguments{
\item{puddlr}{A puddlr object}
}
\value{
puddlr object with pca reduction
}
\description{
Runs PCA for linear dimensionality reduction on the predictors of a 
normalized puddlr object
}
